I corrected and improved the angles of the drawing. I also defined some functions and made it work as well.
What I changed
I changed my code inside the mini_mask to improve it and make it look
better. I also changed the code so there are defined functions and made
the functions work as well.

Why I changed it
I did it so it looks better and I could import it if there is a need to.